+++ src/lib/isInt.js
@@ -10,9 +10,8 @@ export default function isInt(str, options) {
   // Get the regex to use for testing, based on whether
   // leading zeroes are allowed or not.
   let regex = (
-    options.hasOwnProperty('allow_leading_zeroes') && options.allow_leading_zeroes ?
-    intLeadingZeroes :
-    int
+    options.hasOwnProperty('allow_leading_zeroes') && !options.allow_leading_zeroes ?
+    int : intLeadingZeroes
   );
 
   // Check min/max