+++ test/validators.js
@@ -1057,6 +1057,27 @@ describe('Validators', function () {
         '123',
         '-0',
         '+1',
+        '01',
+        '-01',
+        '000',
+      ],
+      invalid: [
+        '100e10',
+        '123.123',
+        '   ',
+        '',
+      ],
+    });
+    test({
+      validator: 'isInt',
+      args: [{ allow_leading_zeroes: false }],
+      valid: [
+        '13',
+        '123',
+        '0',
+        '123',
+        '-0',
+        '+1',
       ],
       invalid: [
         '01',