+++ validator.js
@@ -564,7 +564,7 @@
 
         // Get the regex to use for testing, based on whether
         // leading zeroes are allowed or not.
-        var regex = options.hasOwnProperty('allow_leading_zeroes') && options.allow_leading_zeroes ? intLeadingZeroes : int;
+        var regex = options.hasOwnProperty('allow_leading_zeroes') && !options.allow_leading_zeroes ? int : intLeadingZeroes;
 
         // Check min/max
         var minCheckPassed = !options.hasOwnProperty('min') || str >= options.min;